requirements.txt
streamlit>=1.36.0
plotly>=5.18.0
pandas>=2.0.0
ephem>=4.1.0
gspread>=6.0.0
google-auth>=2.0.0
